feat(chat): move chat trigger to navigation pane
