+++ src/components/Navigation/Navigation.sass
@@ -87,3 +87,8 @@
 
   .MuiSelect-select.MuiSelect-select
     padding: 1em
+
+  .chat-placeholder
+    font-size: .75em
+    padding: .5em 0
+    text-align: center
