+++ src/components/Navigation/Navigation.test.js
@@ -20,6 +20,7 @@ beforeEach(() => {
         dayCount: 0,
         farmName: '',
         handleActivePlayerButtonClick: noop,
+        handleChatRoomOpenStateChange: noop,
         handleClickDialogViewButton: noop,
         handleCloseDialogView: noop,
         handleDialogViewExited: noop,
@@ -30,6 +31,7 @@ beforeEach(() => {
         inventory: [],
         inventoryLimit: INFINITE_STORAGE_LIMIT,
         itemsSold: {},
+        isChatAvailable: false,
         isDialogViewOpen: false,
         isOnline: false,
         stageFocus: stageFocusType.FIELD,
