+++ src/components/OnlinePeersView/OnlinePeersView.sass
@@ -1,8 +1,1 @@
 .OnlinePeersView
-  display: flex
-  flex-direction: column
-
-  button
-    font-size: 1.3em
-    min-width: 56px
-    margin: 0.3em auto
